task1/requirements.txt
asgiref==3.11.1
coverage==7.13.4
Django==5.2.11
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
